+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can now view compiled Twine games natively in VS Code.
 ### Added
 
 -   You can run compiled Twine games directly within VS Code.
+-   Live reload updates the running game on a successful build.
 -   Build system: You can now include files without bundling them into the `.html` file.
 -   Build system: New `Watch` task continuously builds the story whenever source code is changed.